# packages/server/src/openapi.yaml
openapi: 3.0.2
info:
  title: Statshammer
  version: 1.0.0

paths:
  /status:
    get:
      operationId: getStatus
      summary: Get Status
      responses:
        200:
          $ref: "#/components/responses/StatusResponse"

  /api/abilities:
    get:
      operationId: getAbilities
      summary: Fetch the list of abilities for the weapon profiles and target
      responses:
        200:
          $ref: "#/components/responses/AbilitiesResponse"

  /api/compare:
    post:
      operationId: getCompare
      summary: Compare the average damage outputs of the given units
      requestBody:
        $ref: "#/components/requestBodies/CompareBody"
      responses:
        200:
          $ref: "#/components/responses/CompareResponse"

  /api/simulate:
    post:
      operationId: getSimulate
      summary: Perform a number of damage simulations for the given units
      requestBody:
        $ref: "#/components/requestBodies/CompareBody"
      parameters:
        - $ref: "#/components/parameters/limit"
      responses:
        200:
          $ref: "#/components/responses/SimulationResponse"

components:
  requestBodies:
    CompareBody:
      content:
        application/json:
          schema:
            type: object
            properties:
              units:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
              target:
                $ref: "#/components/schemas/Target"
            required:
              - units

  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100000
        default: 5000

  responses:
    StatusResponse:
      description: Get the status of the server
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
              version:
                type: string
            required:
              - status
              - version

    AbilitiesResponse:
      description: The Returned list of available abilities
      content:
        application/json:
          schema:
            type: object
            properties:
              weapon:
                type: array
                items:
                  $ref: "#/components/schemas/AbilityDefinition"
              target:
                type: array
                items:
                  $ref: "#/components/schemas/AbilityDefinition"
            required:
              - weapon
              - target

    CompareResponse:
      description: Compare the average damage outputs of the given units
      content:
        application/json:
          schema:
            type: object
            properties:
              units:
                type: object
                additionalProperties:
                  type: string
              results:
                type: array
                items:
                  $ref: "#/components/schemas/AverageDamageResult"
            required:
              - units
              - results

    SimulationResponse:
      description: Perform a number of damage simulations for the given units
      content:
        application/json:
          schema:
            type: object
            properties:
              units:
                type: object
                additionalProperties:
                  type: string
              results:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationResult"
            required:
              - units
              - results

  schemas:
    Unit:
      type: object
      properties:
        id:
          type: string
          example: unit-1
        name:
          type: string
          example: Unit 1
        weapons:
          type: array
          items:
            $ref: "#/components/schemas/Weapon"
      required:
        - id
        - name
        - weapons

    Weapon:
      type: object
      properties:
        num_models:
          type: integer
          example: 5
        attacks:
          anyOf:
            - type: integer
            - type: string
          example: 2
        to_hit:
          type: integer
          example: 3
        to_wound:
          type: integer
          example: 4
        rend:
          type: integer
          example: 0
        damage:
          anyOf:
            - type: integer
            - type: string
          example: 1
        abilities:
          type: array
          items:
            $ref: "#/components/schemas/Ability"
      required:
        - num_models
        - attacks
        - to_hit
        - to_wound
        - rend
        - damage

    Target:
      type: object
      properties:
        abilities:
          type: array
          items:
            $ref: "#/components/schemas/Ability"
      example:
        abilities:
          - type: TARGET_REROLL

    Ability:
      type: object
      properties:
        type:
          type: string
          example: REROLL_ONES
        options:
          type: object
          properties:
            characteristic:
              type: string
              example: to_hit
      required:
        - type
        - options

    AbilityDefinition:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        options:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AbilityDefinitionOption"
      required:
        - id
        - name
        - description
        - options

    AbilityDefinitionOption:
      oneOf:
        - $ref: "#/components/schemas/ChoiceOption"
        - $ref: "#/components/schemas/NumberOption"
        - $ref: "#/components/schemas/RollOption"
        - $ref: "#/components/schemas/BooleanOption"
      discriminator:
        propertyName: type

    ChoiceOption:
      type: object
      properties:
        type:
          type: string
          enum:
            - choice
        items:
          type: array
          items:
            type: string
        default:
          type: string
      required:
        - type
        - items

    NumberOption:
      type: object
      properties:
        type:
          type: string
          enum:
            - number
        allow_dice:
          type: boolean
        default:
          type: number
      required:
        - type

    RollOption:
      type: object
      properties:
        type:
          type: string
          enum:
            - roll
        allow_ones:
          type: boolean
        default:
          type: number
      required:
        - type

    BooleanOption:
      type: object
      properties:
        type:
          type: string
          enum:
            - boolean
        default:
          type: boolean
      required:
        - type

    AverageDamageResult:
      type: object
      properties:
        save:
          type: number
        displaySave:
          type: string
        values:
          type: object
          additionalProperties:
            type: number
      required:
        - save
        - displaySave
        - values

    SimulationResult:
      type: object
      properties:
        save:
          type: number
        displaySave:
          type: string
        discrete:
          type: array
          items:
            $ref: "#/components/schemas/ProbabilityData"
        cumulative:
          type: array
          items:
            $ref: "#/components/schemas/ProbabilityData"
        metrics:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Metric"
      required:
        - save
        - displaySave
        - discrete
        - cumulative
        - metrics

    ProbabilityData:
      type: object
      properties:
        damage:
          type: number
      additionalProperties:
        type: number

    Metric:
      type: object
      properties:
        max:
          type: number
        average:
          type: number
      required:
        - max
        - average